+++ packages/generator-sketch/src/utils.js
@@ -102,9 +102,48 @@ function setResizingConstraints(symbol, ...predicateActionPairs) {
   }
 }
 
+const searchObject = function (
+  object,
+  matchCallback,
+  currentPath,
+  result,
+  searched
+) {
+  currentPath = currentPath || '';
+  result = result || [];
+  searched = searched || [];
+  if (searched.indexOf(object) !== -1 && object === Object(object)) {
+    return;
+  }
+  searched.push(object);
+  if (matchCallback(object)) {
+    result.push({ path: currentPath, value: object });
+  }
+  try {
+    if (object === Object(object)) {
+      for (var property in object) {
+        if (property.indexOf('$') !== 0) {
+          searchObject(
+            object[property],
+            matchCallback,
+            currentPath + '.' + property,
+            result,
+            searched
+          );
+        }
+      }
+    }
+  } catch (e) {
+    console.log(object);
+    throw e;
+  }
+  return result;
+};
+
 module.exports = {
   findLayer,
   findLayers,
+  searchObject,
   printSymbolStructure,
   setResizingConstraints,
   Resize,